--- Dockerfile-Management
@@ -1,18 +1,0 @@
-# Use an official Python runtime as a parent image
-FROM python:3.8-slim
-
-# Set the working directory to /app
-WORKDIR /app
-
-# Copy the current directory contents into the container at /app
-COPY . /app
-
-# Install any needed packages specified in requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Make the script executable
-RUN chmod +x ./scripts/firewall_setup.sh
-
-ENV FLASK_APP=management_app.py
-
-CMD ["flask", "run", "--host=0.0.0.0"]
